Separate stack-switching assembly from timer ISR
There is now an assembly function 'schedule', that you can call from anywhere
in the kernel safely.  In running threads, this will return to an ISR of some
sort; in new threads, it will return to the very short function 'thread_start'
(consisting of a single 'iret').  We now fake two stack frames in new threads,
but there's still less stuff there.