Add keyboard driver
There's a new syscall 'getch', which gets a character from the keyboard
driver's buffer.  If the buffer is empty, it sleeps the calling thread.  The
next time the keyboard driver gets a key, it'll wake a sleeping thread.

The strategy for which thread it wakes is bad.  There's also no way of
distinguishing why threads are sleeping, which is probably bad.