Set up paging more carefully
- Protect the kernel from access by userspace
- Don't map the null page
- Don't map anything that we're not actually using

Some more details of the page layout:

There's a two level page table.  Each thread has an upper-level table
associated with it, and some lower-level ones.  The FIRST lower-level page
table covers 0x00000000-0x003FFFFF, and is restricted to kernel-only access.
The LAST lower level page table is exactly the same as the upper-level page
table: by doing this, the entire page table is accessible in memory at
0xFFC00000-0xFFFFFFF.

Because the first lower-level page table is the same for every thread, it's
shared.