Stop faffing about with weird virtual memory tricks
The kernel now has the first 512 GB of memory identity-mapped.  memory-map.h
is much barer now, as we're dynamically allocating all the things that used to
be fixed there.  The shared page table is also simpler.

This clears the way for using a memory allocator that can write metadata to
free pages.