Run the binaries provided as multiboot modules.
You can pass ELFs to ./scripts/TEST-QEMU.sh, which passes them to the
bootloader, which (somewhat carefully) passes them to the kernel, which runs
them.  This makes it easier to test one executable at a time, in case you
don't like the "I'M REALLY ANNOYING" program.

The code in the bootloader which moves the module around is subtle and
probably not entirely correct.  If something breaks when starting 2MB of
executables, it might be the reason.