Power machine off when no userland threads remain.
Turning the machine off properly requires APM or ACPI, which are too old and
too complicated for me right now.  But, QEMU also has a special ISA device
that turns the system off.  Let's use that for now.

The main use of this is for testing: once all of your test threads have
exited, QEMU will exit.