Store schedulable threads in a queue
Now we don't have to iterate over the array of threads---indeed, they don't
need to be in an array at all.  As soon as we have 'malloc', the array is
entirely unnecessary.  (Then thread_state is also unnecessary---at this point,
it's just keeping track of the allocation state.)

Be careful about blindly calling 'schedule' now---if you call schedule without
putting yourself in some sort of wait queue, you will never, ever be
rescheduled.