Simplify module manipulation in bootloader
Before, it made fairly week assumptions, but used dubious untested code to
make it work.  Now it has a very simple, strong assumption: All of the
multiboot things are below the address that the kernel will be loaded at.

The advantage is that now modules don't need to be moved around, and their
names will survive.

This can be revisited if the kernel gets larger, but it's definitely
sufficient for now.