Save all registers during syscalls.
Furthermore, save a pointer to the saved registers in the TCB, so that syscall
handlers can easily inspect and modify them.  This makes altering the return
value easier in some cases because you can alter another thread's registers, as
well as your own.  (This happens notably in 'fork' and 'kbd_isr'; the former
was done during this commit because refactoring was already needed there;
kbd_isr will be next.)