Refactor kbd.c in preparation for one-stack
Now no work is ever done after the call to 'schedule'.  This incidentally fixes
bug #16, because now the ordering of tasks waiting on keyboard input is
respected exactly.

Because 'getch' sometimes yields, it won't be able to use a trampoline in
syscall.c once one-stack lands; so make that switch now.

Remove 'wait_event' and 'wake_up'; they are now unused, and the model of
'wait_event' in particular is not compatible with the one-stack refactor.