Refactor ipc.c in preparation for one-stack
Now no work is ever done after the call to 'schedule'.  Because we need to
copy_to_user to the destination's address space, this required adding a hook to
'finish_context_switch' to do the last part of the work.

Because sendrecv sometimes (usually) switches threads, it won't be able to use
a trampoline in syscall.c once one-stack lands; so make that switch now.