Refactor hpet.c in preparation for one-stack
Don't store HPET wait queue on the kernel stack.  The HPET nanosleep wait queue
contained special structs stored in the 'hpet_nanosleep' stack frame of each
sleeping thread; instead, use the standard 'wait_queue' field in the tcb.  This
is necessary to move to a 'single kernel stack' model, and is also probably
easier to understand in general.

Because hpet_nanosleep sometimes (often) yields, it won't be able to use a
trampoline in syscall.c once one-stack lands; so make that switch now.