Respect Intel hardware exception status codes.
Some exceptions on Intel push a status code in addition to the normal 5 values.
(In rough order of importance, #PF, #GP, #DF, #TS, #NP, #SS, and #AC).  The
assembly for the pagefault handler and general protection fault handler (which
also stands in for a couple of other fault-y handlers) didn't understand this;
when they pushed all registers and interpreted their stack as 'struct
all_registers', they were quite incorrect.

The easiest way is to make room in 'struct all_registers' for the status code,
and push 0 if there is no status.  (Popping it off and moving it to a register
would be cleaner, but that would require a scratch register.)