Add userland exception-handling feature.
Threads can choose another thread to be their 'exception handler thread'.  That
thread will receive a message from the kernel upon page faults in the main
thread.  (If they are not ready for that message, the main thread will exit as
usual.)  When the handler thread responds to the main thread, the main thread
will resume... hopefully succesfully.  All arguments in that message are
dropped.